+++ spec/fixtures/projects/coffee/packages/handlebars-runtime/lib/main.js
@@ -1,2 +1,0 @@
-//handlebars-runtime/lib/main
-RUNTIME = "RUNTIME";