+++ spec/fixtures/projects/coffee/packages/handlebars/lib/main.js
@@ -1,0 +1,2 @@
+//handlebars/lib/main
+RUNTIME = "RUNTIME";